修改继承公共Dao和公共Service的Id属性

+++ cpa-dao/src/main/java/cn/kanyun/cpa/dao/itempool/CpaOptionDao.java
@@ -6,7 +6,7 @@ import cn.kanyun.cpa.model.entity.itempool.CpaOption;
 /**
  * Created by Administrator on 2017/6/16.
  */
-public interface CpaOptionDao extends CommonDao<Integer, CpaOption> {
+public interface CpaOptionDao extends CommonDao<Long, CpaOption> {
     public static final String REPOSITORY_NAME = "cn.kanyun.cpa.dao.itempool.impl.CpaOptionDaoImpl";
 
 }
