+++ cpa-dao/src/main/java/cn/kanyun/cpa/dao/itempool/CpaRepertoryDao.java
@@ -6,7 +6,7 @@ import cn.kanyun.cpa.model.entity.itempool.CpaRepertory;
 /**
  * Created by Administrator on 2017/6/16.
  */
-public interface CpaRepertoryDao extends CommonDao<Integer, CpaRepertory> {
+public interface CpaRepertoryDao extends CommonDao<Long, CpaRepertory> {
     public static final String REPOSITORY_NAME = "cn.kanyun.cpa.dao.itempool.impl.CpaRepertoryDaoImpl";
