+++ cpa-dao/src/main/java/cn/kanyun/cpa/dao/user/AttendanceDao.java
@@ -3,7 +3,7 @@ package cn.kanyun.cpa.dao.user;
 import cn.kanyun.cpa.dao.CommonDao;
 import cn.kanyun.cpa.model.entity.user.Attendance;
 
-public interface AttendanceDao extends CommonDao<Integer,Attendance> {
+public interface AttendanceDao extends CommonDao<Long,Attendance> {
     public static final String REPOSITORY_NAME="cn.kanyun.cpa.dao.user.impl.AttendanceDaoImpl";
