+++ cpa-dao/src/main/java/cn/kanyun/cpa/dao/user/UserCollectDao.java
@@ -4,7 +4,7 @@ package cn.kanyun.cpa.dao.user;
 import cn.kanyun.cpa.dao.CommonDao;
 import cn.kanyun.cpa.model.entity.user.UserCollect;
 
-public interface UserCollectDao extends CommonDao<Integer,UserCollect> {
+public interface UserCollectDao extends CommonDao<Long,UserCollect> {
 	public static final String REPOSITORY_NAME="cn.kanyun.cpa.dao.user.impl.UserCollectDaoImpl";
