+++ cpa-dao/src/main/java/cn/kanyun/cpa/dao/user/UserCommentDao.java
@@ -6,7 +6,7 @@ import cn.kanyun.cpa.model.entity.user.UserComment;
 /**
  * Created by KANYUN on 2017/10/24.
  */
-public interface UserCommentDao extends CommonDao<Integer,UserComment> {
+public interface UserCommentDao extends CommonDao<Long,UserComment> {
     public static final String REPOSITORY_NAME="cn.kanyun.cpa.dao.user.impl.UserCommentDaoImpl";
 
 }
