+++ cpa-dao/src/main/java/cn/kanyun/cpa/dao/user/UserDao.java
@@ -4,7 +4,7 @@ package cn.kanyun.cpa.dao.user;
 import cn.kanyun.cpa.dao.CommonDao;
 import cn.kanyun.cpa.model.entity.user.CpaUser;
 
-public interface UserDao extends CommonDao<Integer,CpaUser> {
+public interface UserDao extends CommonDao<Long,CpaUser> {
 	public static final String REPOSITORY_NAME="cn.kanyun.cpa.dao.user.impl.UserDaoImpl";
 
 	/**
