+++ cpa-dao/src/main/java/cn/kanyun/cpa/dao/user/impl/UserCollectDaoImpl.java
@@ -6,7 +6,7 @@ import cn.kanyun.cpa.model.entity.user.UserCollect;
 import org.springframework.stereotype.Repository;
 
 @Repository(UserCollectDao.REPOSITORY_NAME)
-public class UserCollectDaoImpl extends CommonDaoImpl<Integer, UserCollect> implements UserCollectDao {
+public class UserCollectDaoImpl extends CommonDaoImpl<Long, UserCollect> implements UserCollectDao {
     //通过调用父类的构造函数指定clazz值，即实体类的类类型
     public UserCollectDaoImpl() {
         super(UserCollect.class);
