+++ cpa-dao/src/main/java/cn/kanyun/cpa/dao/user/impl/UserCommentDaoImpl.java
@@ -6,7 +6,7 @@ import cn.kanyun.cpa.model.entity.user.UserComment;
 import org.springframework.stereotype.Repository;
 
 @Repository(UserCommentDao.REPOSITORY_NAME)
-public class UserCommentDaoImpl extends CommonDaoImpl<Integer, UserComment> implements UserCommentDao {
+public class UserCommentDaoImpl extends CommonDaoImpl<Long, UserComment> implements UserCommentDao {
 
     //通过调用父类的构造函数指定clazz值，即实体类的类类型
     public UserCommentDaoImpl() {
