+++ cpa-service/src/main/java/cn/kanyun/cpa/service/itempool/CpaOptionService.java
@@ -6,6 +6,6 @@ import cn.kanyun.cpa.service.CommonService;
 /**
  * Created by Administrator on 2017/6/16.
  */
-public interface CpaOptionService extends CommonService<Integer,CpaOption> {
+public interface CpaOptionService extends CommonService<Long,CpaOption> {
     public static final String SERVICE_NAME="cn.kanyun.cpa.dao.itempool.impl.CpaOptionServiceImpl";
 }
