+++ cpa-service/src/main/java/cn/kanyun/cpa/service/user/AnswerRecordService.java
@@ -3,7 +3,7 @@ package cn.kanyun.cpa.service.user;
 import cn.kanyun.cpa.model.entity.user.AnswerRecord;
 import cn.kanyun.cpa.service.CommonService;
 
-public interface AnswerRecordService extends CommonService<Integer, AnswerRecord> {
+public interface AnswerRecordService extends CommonService<Long, AnswerRecord> {
     public static final String SERVICE_NAME = "cn.kanyun.cpa.service.user.impl.AnswerRecordServiceImpl";
 
     /**
