+++ cpa-service/src/main/java/cn/kanyun/cpa/service/user/AttendanceService.java
@@ -5,7 +5,7 @@ import cn.kanyun.cpa.model.entity.user.Attendance;
 import cn.kanyun.cpa.model.entity.user.CpaUser;
 import cn.kanyun.cpa.service.CommonService;
 
-public interface AttendanceService extends CommonService<Integer,Attendance> {
+public interface AttendanceService extends CommonService<Long,Attendance> {
     public static final String SERVICE_NAME = "cn.kanyun.cpa.service.user.impl.AttendanceServiceImpl";
 
     /**
